# swagger.json
{
  "info": {
    "version": "1.0.0",
    "title": "MSPR-API",
    "description": "API du projet MSPR",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "swagger": "2.0",
...
  "securityDefinitions": {
    "JWT": {
      "type": "apiKey",
      "name": "Json Web Token",
      "description": "JWT is retrieved when the user logs in",
      "in": "header"
    }
  },
  "definitions": {
    "ApiResponse": {
      "properties": {
        "success": { "type": "boolean" },
        "code": { "type": "integer" },
        "message": { "type": "string" },
        "data": { "type": "object" }
      },
      "type": "object"
    },
...
  },
  "paths": {
...
    "/me/promotions": {
      "get": {
        "summary": "Get all promotions of connected user",
        "tags": ["Connected user"],
        "description": "",
        "security": [
          {
            "JWT": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of promotions retrieved when scan of qr code",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/me/promotions/{id}": {
      "delete": {
        "summary": "Delete promotion linked to connected user",
        "security": [
          { "JWT": [] }
        ],
        "description": "",
        "tags": ["Connected user"],
        "responses": {
          "200": {
            "description": "Promotion deleted successfully",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          },
          "400": {
            "description": "Invalid request. Promotion id is not valid?"
          },
          "403": {
            "description": "You're not authorized to delete this promotion"
          }
        }
      }
    }
  }
}